requirements.txt
africastalking==1.2.8
aiohttp==3.9.0
aiosignal==1.3.1
amqp==5.1.1
aniso8601==7.0.0
annotated-types==0.5.0
api==0.0.7
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==24.4.2
Automat==22.10.0
aws-xray-sdk==2.13.0
beautifulsoup4==4.12.0
billiard==4.1.0
black==23.1.0
blinker==1.7.0
boto3==1.28.8
botocore==1.31.8
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
cfn-flip==1.3.0
channels==2.4.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clipboard==0.0.4
colorama==0.4.6
compose_x_common==1.4.7
compose_x_render==1.1.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==39.0.2
cssselect2==0.7.0
daphne==2.5.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
Django==3.2.23
django-allauth==0.53.1
django-appconf==1.0.5
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.3.0
django-environ==0.10.0
django-extensions==3.2.3


django-phonenumber-field==7.1.0
django-prometheus==1.0.15


django-storages==1.14.2
django-timezone-field==7.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1



gunicorn==20.1.0
honeycomb-opentelemetry==0.5.0b0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.0
importlib_resources==6.4.0

nose==1.3.7
oauthlib==3.2.2
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.43b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
oscrypto==1.3.0
packaging==23.0
pathspec==0.11.1
phonenumbers==8.13.14
pilkit==2.0
pillow==10.2.0
pip-api==0.0.30
pip-review==1.3.0
platformdirs==3.1.1
pluggy==1.0.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psycopg2==2.9.9
psycopg2-binary==2.9.9

PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.0.9
pypdf==3.6.0
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.2.2
pytest-django==4.5.2
pytest-reqs==0.2.1
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytweening==1.0.7



requests==2.28.2
requests-oauthlib==1.3.1
responses==0.23.1
retry2==0.9.5
rpds-py==0.18.1
rq==1.15.1
Rx==1.6.3
s3transfer==0.6.1
schema==0.7.7
service-identity==24.1.0
singledispatch==4.1.0
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
svglib==1.5.1
tabulate==0.8.10
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
troposphere==4.8.0
troposphere-awscommunity-applicationautoscaling-scheduledaction==0.1.1
Twisted==24.7.0
txaio==23.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.4.0
wrapt==1.16.0
xhtml2pdf==0.2.9
xmltodict==0.13.0
yarl==1.9.3
zipp==3.15.0
zope.interface==7.0.1
